pyproject.toml
[tool]

[tool.poetry]
name = "Django Mason Starter"
version = "0.0.1"
description = "Django Mason Starter"
authors = ["David Dahan <david.dahan3@gmail.com>"]

[tool.poetry.dependencies]

##########################################################################################
# Main
##########################################################################################
python = "3.11.6"
Django = "4.*"       # https://github.com/django/django
django-environ = "*" # https://github.com/joke2k/django-environ

##########################################################################################
# Security
##########################################################################################
django-cors-headers = "*" # https://github.com/adamchainz/django-cors-headers

##########################################################################################
# API
##########################################################################################
orjson = "*"       # https://github.com/ijl/orjson
django-ninja = "*" # https://github.com/vitalik/django-ninja

##########################################################################################
# Background tasks
##########################################################################################
celery = { extras = [
    "redis",
], version = "*" } # https://github.com/celery/celery
django-celery-results = "*" # https://github.com/celery/django-celery-results
django-celery-beat = "*" # https://github.com/celery/django-celery-beat

##########################################################################################
# Database
##########################################################################################
dj-database-url = "*" # https://github.com/jazzband/dj-database-url
psycopg = { extras = [
    "binary",
], version = "*" } # https://github.com/psycopg/psycopg
# boto3 = "*"           # https://github.com/boto/boto3
# django-storages = "*" # https://github.com/jschneier/django-storages/

##########################################################################################
# Additional fields, tags, etc.
##########################################################################################
django-extensions = "*" # https://github.com/django-extensions/django-extensions
django-positions = "*" # https://github.com/jpwatts/django-positions
django-phonenumber-field = { extras = [
    "phonenumbers",
], version = "*" } # https://github.com/stefanfoulis/django-phonenumber-field
django-safedelete = "*" # https://github.com/makinacorpus/django-safedelete

##########################################################################################
# DX & Debug
##########################################################################################
django-debug-toolbar = "*" # https://github.com/jazzband/django-debug-toolbar/
django-rich = "*"          # https://github.com/adamchainz/django-rich
ruff = "*"                 # https://github.com/charliermarsh/ruff
ipdb = "*"                 # https://github.com/gotcha/ipdb

[tool.black]
target-version = ["py311"]

[tool.ruff]
line-length = 90
target-version = "py311"
extend-exclude = ["migrations", "__pycache__"]
per-file-ignores = {}

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.isort]
# 🫤 each module must be described
known-first-party = ["core", "badges", "mailing", "profiles"]
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]
[tool.ruff.isort.sections]
"django" = ["django"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
